fix retainerId type on InventoyItem numeric deserialization FromNumeric()

+++ Models/InventoryItem.cs
@@ -1148,7 +1148,7 @@ namespace CriticalCommonLib.Models
                 SortedContainer = (InventoryType)serializedItem[19],
                 SortedCategory = (InventoryCategory)serializedItem[20],
                 SortedSlotIndex = (int)serializedItem[21],
-                RetainerId = (uint)serializedItem[22],
+                RetainerId = serializedItem[22],
                 RetainerMarketPrice = (uint)serializedItem[23],
                 GearSets = gearSets,
             };
